Add fixtures and tests

+++ mappers/stakingseqmapper/staking_seq_mapper.go
@@ -24,7 +24,7 @@ func ToSequence(stateSyncable syncable.Model) (*stakingseq.Model, errors.Applica
 
 		TotalSupply:         types.NewQuantityFromBytes(stateData.Staking.TotalSupply),
 		CommonPool:          types.NewQuantityFromBytes(stateData.Staking.CommonPool),
-		DebondingInterval:   uint64(stateData.Staking.Parameters.DebondingInterval),
+		DebondingInterval:   stateData.Staking.Parameters.DebondingInterval,
 		MinDelegationAmount: types.NewQuantityFromBytes(stateData.Staking.Parameters.MinDelegationAmount),
 	}
